Add `useOutputAsError` option.
The `useOutputAsError` option uses the compiler output as the error message
passed to webpack. By default, a simple "PureScript compilation has failed."
error is passed to webpack and the compiler output is written to stderr. This
default behavior is undesirable if the compiler output is needed by webpack or
webpack plugins.

Relates to ethul/purescript-webpack-plugin#18 and ethul/purs-loader#40