Push output and errors to the webpack compilation
The PureScript output is now added to the warnings array in the webpack
compilation. Also, any PureScript error is added to the errors array in
the webpack compilation.

This allows for tools such as `psa` to transform the error/output, and
it also allows for the browser to display errors during hot reloading.

Resolves #18